requirements.txt
# Web
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Config
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# CV processing
python-docx>=1.0.0
pdfplumber>=0.10.0
